BAM inactivity timeout in start-up script
Configuring inactivity timeout in WLAN start-up script. Right now
timeout is configured to be 2 seconds. This is the timeout used
by HBM to configure inactivity timeout. When there is no activity
on HSIC BAM for 2 seconds then HBM driver calls the inactivity
call back. This is the indication used for Suspend operation.

Change-Id: Iecc95474d6e5401d6f22f9b118f76e784e46d07b