sstate fixes
lk: update do_deploy task

This fixes the task ordering to enable sstate use.

define LIBGCC, lk was unable to determine the location
when the toolchain comes from sstate.

compat: install header to install location

The sstate package was missing this header since it was
being installed to a different location.

Change-Id: Ia53dffea0389f7275a9605f20d822462986085c6