bluetooth: Check TTY Serial buffer prior to reading data
select() is done as part of the read_vs_hci_event(), but not as
part of the read_hci_event() function. As part of the
select() call, we wait 3 secs for data to be available on the
TTY buffers. If data is readily available, then we return
immediately going on to read it or else select times out. It is a
safe practice for user-space programs to query the availability of
data on the TTY buffers prior to reading from it.

Change-Id: I89246e20fa9cdf5146fb61f9a95ea0ea03dcbde9