rtsp-alg: remove custom configure step.
This configure step compiles rtsp-alg, remove it to use the default
which does not compile rtsp-alg. The modpost permission denied error
happens when the modpost script has not finished being built in the
make_script step, which runs in a parallel to configure step, and the
custom configure step tries to use modpost.

The default configure step does not try to compile rtsp-alg, so the
modpost error will not be seen.

Conflicts:
        recipes/rtsp-alg/rtsp-alg_3.3.bb

Change-Id: Ief1fe6ef3c2a1be14f20f7e07b8d4d38b98c61b8