Avoid ubiattch and mount of modem nand-partition
In case OTA is interrupted while modem is being upgraded,
device boots in recovery mode and the script tries to
ubiattach & mount modem partition again.
However, if modem upgrade was incomplete due to interruption in
previous run, then there exists no valid ubi-filesystem on the
partition. In this case, either ubiattach would fail or the volume-node
(i.e. /dev/ubix_y) will never be created and the script waits
indefinitely waiting for the same.

Avoid ubiattach/mount operation for modem partition. This is
feasible because modem is upgraded as raw partition with no
assumption about the filesystem that resides on it, thus avoiding need
of mounting the partition.

Change-Id: I320f32b6ee554844503bb13dbf1b1e42365e636a