CVE-2014-9984: Provide correct buffer length to netgroup queries in nscd (BZ#16695)
[cherry picked from commit: c44496df2f090a56d3bf75df930592dac6bba46f]

The buffer to query netgroup entries is allocated sufficient space for
the netgroup entries and the key to be appended at the end, but it
sends in an incorrect available length to the NSS netgroup query
functions, resulting in overflow of the buffer in some special cases.
The fix here is to factor in the key length when sending the available
buffer and buffer length to the query functions.

Change-Id: I9408938e93bea44b664022a462a8dfbcc1077918
Signed-off-by: Akshaya <akshayab@codeaurora.org>