mdm9607: meta-msm: comment the wpa-supplicant EAP_PROXY qmi
Comment the wpa-suplicant EAP_PROXY qmi for the compatibility
between old qmi and new wpa-supplicant, because some APIs which
new wpa-supplicant called in eap_proxy qmi are not included in
the old qmi in kernel.
But the eap proxy qmi is not used for the wlan functions in
wpa-supplicant, so disabled wpa-suplicant EAP_PROXY qmi.

Change-Id: I15e1c8d6a8cfa7b23b17f7c5618421e4a64ef16a
Signed-off-by: Ke Huang <keh@codeaurora.org>